src/requirements.txt
Django==4.2.5
Pillow==10.0.0
django-filter==23.2
djangorestframework==3.14.0
redis==5.0.0
channels==4.0.0
channels-redis==4.1.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.2.0
